+++ Thunderstore/CHANGELOG.md
@@ -6,3 +6,9 @@ EARLY ACCESS
 
 ## 1.0.2
 - Fixed a problem with custom interior that have special doors which prevented foxy from destroying them. For now, he will go through them until I find a better fix!
+
+## 1.1.2
+- Enhanced Audio data with the help of Spinmaster, <-- [this guy] is a legend 
+- Added Configurations for modifying difficulty speed and speed
+- Added configurations to modify rarity
+- Added configuration for MISC stuff